requirements.txt
filelock==3.19.1
fsspec==2025.10.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
pandas==2.3.3
pillow==11.3.0
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
six==1.17.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2026.1
PyYAML>=6.0
timm>=0.9.16
transformers>=4.40.0
gdown>=5.2.0
